refactor: move .test.ts files (1)
