+++ app/routes/users/signout.test.ts
@@ -1,8 +1,8 @@
 import { tinyassert } from "@hiogawa/utils";
 import { describe, expect, it } from "vitest";
-import { testLoader, useUser } from "../../../misc/test-helper";
-import { getResponseSession } from "../../../utils/session-utils";
-import { action } from "../signout";
+import { testLoader, useUser } from "../../misc/test-helper";
+import { getResponseSession } from "../../utils/session-utils";
+import { action } from "./signout";
 
 describe("signout.action", () => {
   const { signin } = useUser({ seed: __filename });
